+++ src/Business/Grand.Business.Checkout/Commands/Handlers/Orders/PartiallyRefundOfflineCommandHandler.cs
@@ -58,7 +58,7 @@ namespace Grand.Business.Checkout.Commands.Handlers.Orders
                 throw new ArgumentNullException(nameof(order));
 
             //total amount refunded
-            decimal totalAmountRefunded = order.RefundedAmount + amountToRefund;
+            double totalAmountRefunded = order.RefundedAmount + amountToRefund;
 
             //update order info
             order.RefundedAmount = totalAmountRefunded;
