+++ src/Business/Grand.Business.Checkout/Queries/Models/Orders/CanPartiallyRefundOfflineQuery.cs
@@ -6,6 +6,6 @@ namespace Grand.Business.Checkout.Queries.Models.Orders
     public class CanPartiallyRefundOfflineQuery : IRequest<bool>
     {
         public PaymentTransaction PaymentTransaction { get; set; }
-        public decimal AmountToRefund { get; set; }
+        public double AmountToRefund { get; set; }
     }
 }
